requirements.txt
PyMySQL == 0.10.1
PyYAML == 5.3.1
pymongo == 3.11.2